pyproject.toml
[tool.poetry]
name = "textworks-python"
version = "0.1.0"
description = ""
authors = ["Peyton Walters <pawalt@hey.com>"]

[tool.poetry.dependencies]
python = "^3.10"
transformers = "^4.37.2"
httpx = "^0.26.0"
python-dotenv = "^1.0.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
